+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.1.3] - 2023-05-29
+## [1.1.3] - 2023-05-30
 * Added default and parametrized constructors to `BWidgets::Support`
 * Fix center `BWidgets::RadialMeter`
 * Add class `BWidgets::ValueVisualizable`
+* Add parametrized overrides of `getExtends()`, `getWidth()`, and `getHeight()` 
+in `BWidgets::Label` to get Widget extends for a not yet applied text string
 
 
 ## [1.1.2] - 2023-05-27